Use attrib arrays in Cg.
Attempt to modernize by avoiding glVertexPointer, etc if possible, so
that CG *might* work in a modern GL setting.

This appears to work fine on PC, but it is unknown if this will work
on PS3s CG implementation. If it works, it might be possible to strip
down RGL more by avoiding glVertexPointer and that jazz.

If not, this should be reverted.