(MSVC 2012) Initial MSVC 2012 solution
Just need to sort out the linking errors - we'll need to create a
Direct3D driver that is not external for PC and do away with the
'external' one which is really an unacceptable extra dependency for such
a popular graphics API. Also, look if the WGL context file will even
work. Added some TODOs too.