Update AUTHORS
Added meancoot to authors - if he wants his real name in the credits he can go ahead and make a pull request with the necessary changes to this file. If not, we'll leave it at the nick.

The same applies to e-mail address - it isn't obligatory but it would be nice to add it.