use the app level documents directory instead of /var/mobile
This is a big change in that it changes the RetroArch system path where
it saves settings, game saves, etc. Using /var/mobile meant that
RetroArch only ran on jailbroken devices. This was the only thing
requiring a jailbreak.

Now not only will it work on vanilla devices, but all settings etc are
neating found in the app's documents directory.