retrolaunch: Change core selection in retrolaunch
This changed makes it so that retrolaunch will choose cores based on a
preference configuration called cores.conf.

This means that different systems can select what cores they prefer to
use when running certain games. Games that don't work on specific cores
can be marked in launch.conf with !core_name and retrolaunch will choose
the next core for that system.

Also launch.conf is now optional meaning that users don't need the cddb
and the dat files to be able to use core auto-selection. They just need
cores.conf.

Signed-off-by: Saggi Mizrahi <ficoos@gmail.com>