video_lima: adjust aspect ratio depending on frame dimensions
With this the emulator framebuffer is no longer stretched to the entire
screen. However it's still not entirely correct. For example the aspect
ratio value provided by the frontend is not used yet.
Also I noticed this behaviour: When using the game "Secret of Mana 2"
(unofficial english translation) on the snes9x-next core, the game
normally uses a framebuffer of dimension 256x224. But when rendering
text windows, it increases resolution to 512x224. With the current
behaviour this alters aspect ratio and let's the image "explode"
when text boxes are displayed, only to "implode" again when they close.
This should not happen! (Needs further investigation...)