Add rudimentary Core Manager - pulls index listing from the buildbot and lets you download cores to the core directory.
Current limitations and what still needs to be addressed:
* Core is downloaded to core directory still packaged in a zip
file. Will need manual unzipping by the user. We should have
a setting for automating this.
* We should have some sane defaults for the buildbot URL on
specific platforms. For instance, Android, Windows, etc.
* We should present a fancier name listing than what we have
now eventually where it gets the names from the core info files
* Should we do comparisons between the currently installed core
and the one we're about to download now? A CRC check or some
different means?  Problem I can see is that there's right now
no good way to determine if a core actually needs to be updated
without first downloading it.