Autoconfigure - Change equals to starts with ident
This allows for Bluetooth PS3 controllers (and other bluetooth controllers) to be automatically detected.

In the input_try_autoconfigure_joypad_from_conf function, there is a string comparision between the input's name "params->name" and the configuration's name "ident".  Specifically on Bluetooth controllers, the uuid is appended to the end of the name.  
This enhancement will change the checking for name match from an equals comparision, to a startswith.