Added check to make sure ident was not an empty string.
Added check to make sure ident was not an empty string.

This commit is to fix a bug related to having a autoconfig.cfg file without an ident.
Currently, there are none in the repository that fit this condition, which is why the bug was never triggered in testing. 

This checks to make sure the ident loaded from the config file actually has a value associated with it.
Without this check, the if() statement will always return true, because an empty string is always within another string.