(Linux) Install the desktop entry for everyone in the top-level Makefile
I've also removed the PNG from installation because the icon
specification will cause it to be prefered over the SVG version when
picking the icon.  I have edited the desktop entry to refer to the
icon by basename only, this will allow it to be properly selected
regardless of whether retroarch is installed in /usr or /usr/local

The desktop entry has also been cleaned up to the proper desktop entry
specifications.