Allow GLES2 app to use GLES3 core
The Android app is a GLES2 application (it doesn't have HAVE_OPENGLES3 set when it is compiled). However, it is possible to use the GLES2 Android App to launch GLES3 cores, if the phone supports it. This allows cores to set RETRO_HW_CONTEXT_OPENGLES3 and have the Android app respect that, without having to compile the Android app for GLES3.