Fixes #1868: Added Tracis CI config file
Builds on Linux with gcc and clang, and also on Mac.

Had to configure a more recent version of clang as the default Travis CI
one is a bit outdated and wouldn't work.

Unfortunately I was unable to install Nvidia Cg for Mac as it needs to
run the installer interactively. Still, it builds without Cg support.

It only covers a subset of the architectures, but that's better than
nothing and should catch at least basic errors on pull requests
(requires enabling Travis CI support in the GitHub repository settings).